feat(mine): 完善过磅管理系统功能
- 新增过磅类型字段支持发货和收货分类管理
- 重构过磅记录实体类添加收货单位和物资名称字段
- 优化过磅和回磅业务逻辑增强数据验证和状态检查
- 完善查询功能支持按过磅类型筛选和分页查询
- 更新DTO类适配新的业务字段和数据类型
- 优化商户名称查询和显示功能支持发货和收货单位
- 添加车牌号去重和唯一性校验防止重复过磅
- 完善异常处理和业务验证提升系统稳定性

+++ divine-modules/divine-mine/src/main/java/com/divine/mine/domain/dto/WeightingReturnDto.java
@@ -4,34 +4,25 @@ import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 
 /**
- * @Author: Yisl
- * @Description:
- * @Date: 2026/3/13 18:08
+ * 回磅 DTO
+ *
+ * @author yisl
+ * @date 2026-03-13
  */
 @Data
 public class WeightingReturnDto {
 
-    /**
-     * 主键id
-     */
+    /** 主键id */
     @NotNull(message = "主键id不为空")
     private Long id;
 
-    /**
-     * 总重
-     */
-    private Long totalWeight;
+    /** 总重 */
+    private Double totalWeight;
 
-    /**
-     * 皮重
-     */
-    private Long tareWeight;
+    /** 皮重 */
+    private Double tareWeight;
 
-    /**
-     * 净重
-     */
-    private Long netWeight;
-
-
+    /** 净重(总重-皮重) */
+    private Double netWeight;
 
 }
